+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add support for experimental code generation with the `plugin:` key in `buf.gen.yaml`.
 - Preserve single quotes with `buf format`.
 - Support `junit` format errors with `--error-format`.
+- Include `LICENSE` file in the module on `buf push`.
 
 ## [v1.7.0] - 2022-06-27